requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.11.2
djangorestframework==3.12.2
requests==2.25.1
urllib3==1.26.2
gunicorn==19.7.1
Werkzeug==1.0.1
whitenoise==5.2.0
dj-database-url==0.5.0
psycopg2-binary
django-heroku